// payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface SiteSetting {
  id: number;
  name: string;
  tagline: string;
  welcome: string;
  heroImage?: (number | null) | Media;
  heroImageUrl?: string | null;
  phone: string;
  email: string;
  address: string;
  /**
   * Google Maps embed URL for the contact page.
   */
  mapEmbedUrl?: string | null;
  hours?:
    | {
        label: string;
        value: string;
        id?: string | null;
      }[]
    | null;
  instagram?: string | null;
  facebook?: string | null;
  twitter?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}